requirements.txt
packaging==21.3
Pillow==9.2.0
pip==22.2.2
PyDirectInput==1.0.4
pyparsing==3.0.9
pytesseract==0.3.10
setuptools==65.4.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
xlib==0.21
